+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-05-24 15:41:29
-// Commit: style(perf): harmonize light and dark theme palette for perf
+// Updated at: 2026-05-24 19:24:09
+// Commit: docs(perf): update readme and API contracts for perf
 
 class perfComponentState {
-  final String commitRef = '1399';
-  final String lastMessage = 'style(perf): harmonize light and dark theme palette for perf';
-  final DateTime timestamp = DateTime.parse('2026-05-24T15:41:29+05:00');
+  final String commitRef = '1415';
+  final String lastMessage = 'docs(perf): update readme and API contracts for perf';
+  final DateTime timestamp = DateTime.parse('2026-05-24T19:24:09+05:00');
 }